requirements.txt
Django==1.9.4
decorator==4.0.9
djangorestframework==3.3.3
ipython==4.1.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
simplegeneric==0.8.1
traitlets==4.2.1
wsgiref==0.1.2
